+++ src/AdVision.Contracts/DiscountDto.cs
@@ -1,3 +1,3 @@
 namespace AdVision.Contracts;
 
-public sealed record DiscountDto(Guid Id, string Name, double Percent, decimal MinTotal);
+public sealed record DiscountDto(Guid Id, string Name, decimal Percent, decimal MinTotal);
